backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.6.0
zhipuai==2.1.5.20250825
moviepy>=1.0.3
Pillow==10.1.0
gTTS==2.4.0
edge-tts==6.1.9
nest-asyncio==1.5.8
pydub==0.25.1
aiofiles==23.2.1
python-dotenv==1.0.0
